Asistente_Virtual (Gabriel Herrador)/requirements.txt
google-generativeai
python-dotenv
tk
